Se crea login de administrador, con controlador (no funciona todabia)

+++ backend/controllers/UsuarioController.php
@@ -8,7 +8,7 @@ use app\models\PostSearch;
 use yii\web\Controller;
 use yii\web\NotFoundHttpException;
 use yii\filters\VerbFilter;
-
+use backend\filtros\AdminControl;
 /**
  * UsuarioController implements the CRUD actions for User model.
  */
@@ -32,6 +32,9 @@ class UsuarioController extends Controller
      */
     public function actionIndex()
     {
+        if( !AdminControl::esAdmin(Yii::$app->user->identity->username) )
+            return $this->redirect(Yii::$app->urlManager->createUrl('./../../frontend/web/'));
+
         $searchModel = new PostSearch();
         $dataProvider = $searchModel->search(Yii::$app->request->queryParams);
 
@@ -48,6 +51,9 @@ class UsuarioController extends Controller
      */
     public function actionView($id)
     {
+        if( !AdminControl::esAdmin(Yii::$app->user->identity->username) )
+            return $this->redirect(Yii::$app->urlManager->createUrl('./../../frontend/web/'));
+
         return $this->render('view', [
             'model' => $this->findModel($id),
         ]);
@@ -60,6 +66,9 @@ class UsuarioController extends Controller
      */
     public function actionCreate()
     {
+        if( !AdminControl::esAdmin(Yii::$app->user->identity->username) )
+            return $this->redirect(Yii::$app->urlManager->createUrl('./../../frontend/web/'));
+
         $model = new User();
 
         if ($model->load(Yii::$app->request->post()) && $model->save()) {
@@ -79,6 +88,9 @@ class UsuarioController extends Controller
      */
     public function actionUpdate($id)
     {
+        if( !AdminControl::esAdmin(Yii::$app->user->identity->username) )
+            return $this->redirect(Yii::$app->urlManager->createUrl('./../../frontend/web/'));
+
         $model = $this->findModel($id);
         
         if ($model->load(Yii::$app->request->post()) && $model->save()) {
@@ -101,6 +113,9 @@ class UsuarioController extends Controller
      */
     public function actionDelete($id)
     {
+        if( !AdminControl::esAdmin(Yii::$app->user->identity->username) )
+            return $this->redirect(Yii::$app->urlManager->createUrl('./../../frontend/web/'));
+
         $this->findModel($id)->delete();
 
         return $this->redirect(['index']);
@@ -113,14 +128,13 @@ class UsuarioController extends Controller
      */
     public function actionConfirmar($id)
     {
+        if( !AdminControl::esAdmin(Yii::$app->user->identity->username) )
+            return $this->redirect(Yii::$app->urlManager->createUrl('./../../frontend/web/'));
         
         $model = $this->findModel($id);
         $model->confirm();
 
         return $this->redirect(['view', 'id' => $model->id]);
-
-        
-
     }
 
     /**
